requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
pydantic
loguru